Refactor: Remove unused QueryType and SqlQueryAnalyzer classes to simplify codebase
